requirements.txt
google-generativeai
pandas
python-dotenv
flask
